agents/agent_MCPNotionAgent/requirements.txt
openai
tiktoken
python-dotenv
urllib3
pydantic
aioconsole

aiohttp

mcp